+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 <cite>Release contributors</code>
 - 10 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2026.0.9+author%3Amlytvyn+is%3Apr)
-- 3 PR(s) by [Eugeni Kalenchuk](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2026.0.9+author%3Aekalenchuk+is%3Apr)
+- 5 PR(s) by [Eugeni Kalenchuk](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2026.0.9+author%3Aekalenchuk+is%3Apr)
 
 ### `Project Import` enhancements
 - Enable `externalDependencies.xml` as a restoration entry on project reimport [#1884](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1884)
@@ -21,6 +21,9 @@
 - Display a SAP Commerce version [#1888](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1888)
 - Support deleting a recent project [#1891](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1891)
 - Display a Git branch [#1892](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1892)
+
+### `SAP CX Logging` enhancements
+- Add filtering for custom and remote loggers [#1896](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1896)
 
 ## [2026.0.8]